Update Plugins
Appointments+
Genesis Author Pro
Gravity Forms
Master Slider
mb.YTPlayer for background videos
Meta Slider
Pretty Plugins
Reveal IDs
WP Offload S3
WP Offload S3 - Assets Addon
WP Offload S3 - Enable Media Replace Addon
WP Offload S3 - Meta Slider Addon
WP Offload S3 - WooCommerce Addon
WP Smush Pro

+++ content/plugins/wpmbytplayer/uninstall.php
@@ -3,8 +3,14 @@
 if( ! defined( 'WP_UNINSTALL_PLUGIN' ) )
     exit();
 
-// Uninstall all the mbYTPlayer settings
-delete_option('mbYTPlayer_version');
+if(!$ytppro) {
+    // Uninstall all the mbYTPlayer settings
+    delete_option('mbYTPlayer_version');
+    delete_option('mbYTPlayer_Home_is_active');
+    delete_option('mbYTPlayer_home_video_url');
+    delete_option('mbYTPlayer_home_video_page');
+}
+/*delete_option('mbYTPlayer_version');
 delete_option('mbYTPlayer_Home_is_active');
 delete_option('mbYTPlayer_home_video_url');
-delete_option('mbYTPlayer_home_video_page');
+delete_option('mbYTPlayer_home_video_page');*/
